draft single figs and set up ssh test

+++ Scenarios/t_ssh_dist.py
@@ -20,7 +20,8 @@ def mp_runner(kwargs):
     brd = Path(kwargs.get('local_run_dir'))
     model_name = kwargs.get("model_name")
     base_outputs_dir = Path(kwargs.get('base_outputs_dir'))
-    shutil.rmtree(base_outputs_dir)
+    if base_outputs_dir.exists():
+        shutil.rmtree(base_outputs_dir)
     logfile = base_outputs_dir.joinpath(f'0_{model_name}.log')
     logfile.parent.mkdir(exist_ok=True, parents=True)
     with logfile.open('w') as f:
